refactor(extraction): relocate assembler to app/llm, delegate serialization to render_blocks_to_markdown
